+++ wp-content/themes/beautifultrouble/single-bt_theory.php
@@ -58,40 +58,26 @@ get_header(); ?>
                         }
                 }
                 ?>
+            <?php if ( get_the_excerpt() ) { ?>
                 <div id="in-summary" class="alert alert-info">
-                    <strong>In Summary</strong>
+                    <strong>In Sum</strong>
                     <?php the_excerpt(); ?>
                 </div> 
+            <?php } ?>
+            <?php if( $fields['origins'] ) {
+            ?>
                 <div id="origins">
                     <strong class="origins">Origins:</strong> <?php echo $fields['origins']; ?>
                 </div>
+            <?php
+                }   
+            ?>
             <?php the_content();?>
             <?php the_tags( '<p>Tags: ', ', ', '</p>'); ?>
 <?php endwhile; // end of the loop. ?>
 
         </div><!-- /.span8 -->
         <div id="marginalia" class="fluid-sidebar sidebar span4" role="complementary">
-            <?php if( $fields['key_principle_at_work'] ) {
-                echo '<div id="key-principle" class="alert alert-success">';
-                echo '<strong id="key-principle">Key Principle at work</strong><br />';
-                // Key Principle At Work
-                $principles = $fields['key_principle_at_work'];
-                foreach( $principles as $principle ) {
-                    $related = array_pop( $principle['related_principle'] );
-                    echo '<p class="principle"><b><a href="' . $related->guid . '">' . $related->post_title . '</a></b><br />';
-                    echo $principle['explanation'];
-                    echo '</p>';
-                }
-                echo '</div>';
-            }
-            ?>
-            <?php if( $fields['potential_pitfalls'] ) {
-                // Potential Pitfalls
-                echo '<div class="alert">';
-                echo '<strong id="potential-pitfalls">Potential Pitfalls</strong>';
-                echo '<p class="pitfalls">' . $fields['potential_pitfalls'] . '</p>';
-                echo '</div>';
-            } ?>
             <?php if( $fields['repeating_elements'] ) {
                 if( $further_insights ) {
                 // Further Insights